Refactor to create visitors module
